Final Milestone

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -1,8 +1,11 @@
 package ImageHoster.controller;
 
+
+import ImageHoster.model.Comment;
 import ImageHoster.model.Image;
 import ImageHoster.model.Tag;
 import ImageHoster.model.User;
+import ImageHoster.service.CommentService;
 import ImageHoster.service.ImageService;
 import ImageHoster.service.TagService;
 
@@ -29,6 +32,9 @@ public class ImageController {
     @Autowired
     private TagService tagService;
 
+    @Autowired
+    private CommentService commentService;
+
     //This method displays all the images in the user home page after successful login
     @RequestMapping("images")
     public String getUserImages(Model model) {
@@ -47,10 +53,15 @@ public class ImageController {
     //Also now you need to add the tags of an image in the Model type object
     //Here a list of tags is added in the Model type object
     //this list is then sent to 'images/image.html' file and the tags are displayed
+    // replaced title with id attribute to fetch single record
+    // Comment service implemented to get the comments by image id
     @RequestMapping("/images/{imageId}/{title}")
-    public String showImage(@PathVariable("imageId") Integer id,@PathVariable("title") String title, Model model) {
+    public String showImage(@PathVariable("imageId") Integer id, @PathVariable("title") String title, Model model) {
         Image image = imageService.getImage(id);
+        List<Comment> comments  = commentService.getAllComments(id);
+        model.addAttribute("comments",comments);
         model.addAttribute("image", image);
+        model.addAttribute("title", title);
         model.addAttribute("tags", image.getTags());
         return "images/image";
     }
@@ -94,18 +105,18 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, HttpSession session, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) throws Exception {
 
-       /* String error = "Only the owner of the image can edit the image";
-        model.addAttribute("editError", error);*/
-
+        User user = (User) session.getAttribute("loggeduser");
         Image image = imageService.getImage(imageId);
-        String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
-        model.addAttribute("tags", tags);
-        return "images/edit";
-
-
+        model.addAttribute("tags", image.getTags());
+        if(user.getId()!= image.getUser().getId()){
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("editError", error);
+            return "images/image";
+        }
+            return "images/edit";
     }
 
     //This controller method is called when the request pattern is of type 'images/edit' and also the incoming request is of PUT type
@@ -147,7 +158,16 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, Model model) {
+        User user = (User) session.getAttribute("loggeduser");
+        Image image = imageService.getImage(imageId);
+        model.addAttribute("image", image);
+        model.addAttribute("tags", image.getTags());
+        if (user.getId() != image.getUser().getId()) {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("deleteError",error);
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
